centralize def type
