+++ backends/aoti/runtime/shims/memory.h
@@ -8,66 +8,49 @@
 
 #pragma once
 
-#include <cuda_runtime.h>
-#include <executorch/extension/tensor/tensor.h>
-#include <executorch/runtime/core/error.h>
 #include <torch/csrc/inductor/aoti_torch/generated/c_shim_cuda.h>
 #include <cstdint>
 #include <memory>
 #include <unordered_map>
 #include <unordered_set>
 #include <vector>
+#include "types.h"
 
 namespace executorch {
 namespace backends {
 namespace aoti {
 
-using executorch::runtime::Error;
-using executorch::runtime::etensor::Tensor;
-
 extern "C" {
-
-// Type definitions
-using AOTITensorHandle = Tensor*;
-using AOTIRuntimeError = Error;
-using AOTITorchError = Error;
-
-struct CUDAStreamGuardOpaque {
-  cudaStream_t original_stream;
-  int device_index;
-  cudaEvent_t sync_event;
-};
-using CUDAStreamGuardHandle = CUDAStreamGuardOpaque*;
 
 // Global storage declarations
 extern std::unordered_map<Tensor*, bool> is_tensor_own_memory;
 extern std::unordered_set<std::shared_ptr<Tensor>> tensors;
 
 // Memory-related operations
-AOTITorchError aoti_torch_create_tensor_from_blob_v2(
-    void* data,
-    int64_t ndim,
-    const int64_t* sizes_ptr,
-    const int64_t* strides_ptr,
-    int64_t storage_offset,
-    int32_t dtype,
-    int32_t device_type,
-    int32_t device_index,
-    AOTITensorHandle* ret_new_tensor,
-    int32_t layout,
-    const uint8_t* opaque_metadata,
-    int64_t opaque_metadata_size);
+// AOTITorchError aoti_torch_create_tensor_from_blob_v2(
+//     void* data,
+//     int64_t ndim,
+//     const int64_t* sizes_ptr,
+//     const int64_t* strides_ptr,
+//     int64_t storage_offset,
+//     int32_t dtype,
+//     int32_t device_type,
+//     int32_t device_index,
+//     AOTITensorHandle* ret_new_tensor,
+//     int32_t layout,
+//     const uint8_t* opaque_metadata,
+//     int64_t opaque_metadata_size);
 
-AOTITorchError aoti_torch_create_tensor_from_blob(
-    void* data,
-    int64_t ndim,
-    const int64_t* sizes_ptr,
-    const int64_t* strides_ptr,
-    int64_t storage_offset,
-    int32_t dtype,
-    int32_t device_type,
-    int32_t device_index,
-    AOTITensorHandle* ret_new_tensor);
+// AOTITorchError aoti_torch_create_tensor_from_blob(
+//     void* data,
+//     int64_t ndim,
+//     const int64_t* sizes_ptr,
+//     const int64_t* strides_ptr,
+//     int64_t storage_offset,
+//     int32_t dtype,
+//     int32_t device_type,
+//     int32_t device_index,
+//     AOTITensorHandle* ret_new_tensor);
 
 AOTITorchError aoti_torch_empty_strided(
     int64_t ndim,
