+++ backends/aoti/runtime/shims/utils.h
@@ -8,26 +8,18 @@
 
 #pragma once
 
-#include <executorch/extension/tensor/tensor.h>
-#include <executorch/runtime/core/error.h>
-#include <torch/csrc/inductor/aoti_torch/c/shim.h>
+#include <executorch/runtime/core/exec_aten/exec_aten.h>
 #include <cstdint>
+#include "types.h"
 
 namespace executorch {
 namespace backends {
 namespace aoti {
 
-using executorch::runtime::Error;
-using executorch::runtime::etensor::Tensor;
-
 extern "C" {
 
-// Type definitions
-using AOTITensorHandle = Tensor*;
-using AOTITorchError = Error;
-
 // Utility function for printing tensor information
-void aoti_torch_print_tensor_handle(AtenTensorHandle self, const char* msg);
+void aoti_torch_print_tensor_handle(AOTITensorHandle self, const char* msg);
 
 // Cleanup function for tensor output file (called during backend destruction)
 void cleanup_aoti_tensor_output();
