+++ backends/aoti/runtime/targets.bzl
@@ -14,6 +14,7 @@ def define_common_targets():
             "aoti_model_container.h",
             "shims/memory.h",
             "shims/tensor_attribute.h",
+            "shims/types.h",
             "shims/utils.h",
         ],
         # @lint-ignore BUCKLINT: Avoid `link_whole=True` (https://fburl.com/avoid-link-whole)
